Replaced NULL_NODE, NULL_EDGE etc. with the is_valid check on the key.

+++ is_mesh/is_mesh_key_type.h
@@ -9,7 +9,7 @@ namespace is_mesh
         unsigned int key;
         unsigned int dim;
         
-        Key() : Key(-1,-1)
+        Key() : Key(static_cast<unsigned int>(-1), static_cast<unsigned int>(-1))
         {
             
         }
